Add cmake option VALIDATOR_ONLY to build only the validator, as the name says.
This removes the dependencies on kholidays and ki18n, when all that is
wanted is a validator.

Impact on the API: setRegion/region/interval/nextInterval is not
available when this option is set. setLocation/setTimeZoneId (and
getters) are available but have no effect, maybe I should remove them too.

All the methods for evaluation have been moved to evaluator.cpp.
An alternative to splitting the implementation of internal classes like this
would be to move code into an Evaluator class, but input welcome before
such a big change.