Add the ability to resume parsing after a syntax error
We use this to support wrong additional rule separation. Trying to model
this case in the grammar would introduce so many ambiguities that even GLR
parsing isn't able to handle that anymore. Instead we now take what we
parsed so far, and continue fresh from after the wrong comma.

The same approach should be applicable for the also very common case of
rule separators missing entirely (it technically works, but is a bit too
lenient still).

This results in about 3k additional expressions from the full OSM corpus
to now get accepted.