Deal with slashes wrongly used as time span or rule separators
This works, but comes at a rather heavy price of more than doubling the
ambiguities in the grammar. GLR parsing still seems to hold up with that,
but there is only so much it can handle, so we better don't need more
changes like this.

Overall this only parses another ~200 expressions from the full OSM corpus,
so a fairly limited gain unfortunately.