Fix normalization of Su[-2,-1]
We were generating Su[-2--1] which can't be parsed back (neither by
KOpeningHours nor by most humans, lol).

Looking at the spec I notice that ranges (Su[1-3]) are not in the spec
anyway, only comma-separated values. But I assume this is used out
there, so keep the support for intervals, just not when the second
number is negative.