Generalize the weekday autocorrect
Sep 16-Jun 15: Tu-Fr, Sa, Su [1,3] 09:00-14:00
was normalized as
Sep 16-Jun 15: Tu-Fr, Sa,Su[1,3] 09:00-14:00
with a missing weekday autocorrects: the space before the first "Sa".
This is because the monthday selector was preventing the first two rules
from being collapsed together.

This commit generalizes the weekday autocorrect to accept any additional
selector in the first rule, except of course a time selector.
It adds made-up testcases for the other cases of selectors in the first
rule (a year range, a week range).
I also grouped together all testcases that are mostly about the weekday
autocorrect.