Don't do error recovery between two wrongly ordered wide range selectors
We end up in this path e.g. for wrong date formats, and since all parts
are technically valid, error recovery tried to piece something together
out of those. The current check for things being in the right order is
far from complete, but we can extend that as needed, we aren't accepting
more input than before due to this.