autocorrect wrong use of ';' where ',' was expected
We can detect the common case of
- Mo-Sa 12:00-15:00; 18:00-24:00
+ Mo-Sa 12:00-15:00,18:00-24:00
based on the fact that the second rule has no time selector.

The JS evaluation_tool doesn't do this autocorrect though, and
adds the two time ranges (as if "Mo-Sa 12:00-15:00, Mo-Su 18:00-24:00")
rather than overwriting the first rule with the second as I thought would happen.

It's still very unlikely that this is what the author meant though...

Do the same in case of identical weekdays (and no other selector):
- Mo-Sa 12:00-15:00; Mo-Sa 18:00-24:00
+ Mo-Sa 12:00-15:00,18:00-24:00