Add simplifiedExpression()
This simplifies the expression if possible:
- Mo 08:00-13:00; Tu 08:00-13:00
+ Mo,Tu 08:00-13:00

Doing this inside normalizedExpression() would create many more Osmose
errors than necessary. Better differenciate normalization (which helps
detecting invalid input) and simplification (which is more about setting
the best possible expression if we're modifying it anyway).

This comes from the fact that I got comments on changesets where
I was setting a non-simplified expression. It's all the same to
computers, but humans find the simplified version more readable/logical.