Don't turn 20:00-26:00 into 20:00-02:00
Changing the input (like we do when it's invalid) means Osmose users
are told the input is invalid, which isn't the case for 26:00.

Normalize internally for evaluating, but not for toExpression().

This required adding a bool for pointInTime, since begin=10 end=10
can be either 10:00 or 10:00-10:00 (no longer stored as 10:00-34:00).
It makes things pretty straightforward actually: if there was a '-'
in the input, toExpression() outputs a '-', otherwise it doesn't.