Support 4 digit times to the extend possible
The problem with 4 digit times is the ambiguity with year numbers. To solve
this we now assume anything that would be a valid time outside of the
[2001-2099] range to be a time. That leaves all practically relevant years
valid and still covers the vast majority of 4 digit times found in the OSM
corpus.

This is worth it given how common the 4 digit time mistake is in OSM data.

BUG: 446137