Repeat year and month, when both are specified
This is a followup to the fix for 446224.

Dec 24,Dec 25 can indeed become Dec 24,25
but 2021 Dec 24,25 is incorrect and must become 2021 Dec 24,2021 Dec 25
according to https://openingh.openstreetmap.de/evaluation_tool/

https://wiki.openstreetmap.org/wiki/Key:opening_hours/specification#explain:date_to:daynum
says "A day missing a month is assumed to refer to the last month found
before. So Jan 23-25 is evaluated as Jan 23-Jan 25." but says nothing
about assuming a year too.